Create components to instantiate google maps and markers declaratively
This commit adds two components: `GoogleMap` and `GoogleMarker`. This
was heavily inspired by the react-google-maps library, but without all
the flexibility and with custom interactions on how the map inside
pickup-points-modal works.

These components will allow us to render the map declaratively instead
of using refs and imperative code to handle the rendering of the markers
and side-effects (such as re-centering the map after the user changes
their address), in a much more maintainable fashion.